+++ CHANGELOG.md
@@ -207,3 +207,4 @@
 - 2023-12-12: Pagination deal, Makefile, bettercontact, discover, +2
 - 2023-12-12: Cache leads, Makefile, descend, follow up, +3
 - 2023-12-13: Serde geo, README, lib, chat, +4
+- 2023-12-14: Builder task, frontier, descend, campaign, +3